Extract `FeedCacheSpy` to shared scope to remove duplication
