+++ EssentialFeedTests/Feed Cache/CacheFeedUseCase.swift
@@ -143,44 +143,4 @@ final class CacheFeedUseCase: XCTestCase {
     private func anyNSError() -> NSError {
         NSError(domain: "any error", code: 0)
     }
-    
-    private class FeedStoreSpy: FeedStore {
-        private(set) var receivedMessages = [ReceivedMessage]()
-        
-        enum ReceivedMessage: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        
-        
-        private var deletionCompletions = [DeletionCompletion]()
-        private var insertionCompletions = [InsertionCompletion]()
-        
-        func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-            deletionCompletions.append(completion)
-            receivedMessages.append(.deleteCachedFeed)
-        }
-        
-        func completeDeletionSuccesfully(at index: Int = 0) {
-            deletionCompletions[index](nil)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            deletionCompletions[index](error)
-        }
-        
-        func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-            insertionCompletions.append(completion)
-            receivedMessages.append(.insert(feed, timestamp))
-        }
-        
-        func completeInsertion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSuccessfuly(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-    }
 }
